Refactor Ktor client instance creation logic to an extension function
